refactor(lapis): move field/orderBy converters and cleaners into request.converter
Field.kt and OrderByField.kt mixed request data types with the
converter/cleaner classes that produce them, making the files harder
to scan. Converters and cleaners (FieldConverter, MetadataFieldConverter,
AggregatedFieldConverter, PlainFieldConverter, SequencePositionFieldConverter,
CaseInsensitiveFieldsCleaner, OrderByFieldsCleaner, OrderByFieldConverter)
now live in a dedicated request.converter package; request data types
(Field, PlainField, SequencePositionField, OrderBySpec, OrderByField, Order)
and Jackson deserializers stay in request. Pure move, no behavior change.
