+++ lapis/src/main/kotlin/org/genspectrum/lapis/request/converter/FieldConverter.kt
@@ -1,0 +1,5 @@
+package org.genspectrum.lapis.request.converter
+
+fun interface FieldConverter<T> {
+    fun convert(source: String): T
+}
